mandatory done

+++ philo/src/time/time.c
@@ -6,14 +6,14 @@
 /*   By: itan <itan@student.42kl.edu.my>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/18 03:46:00 by itan              #+#    #+#             */
-/*   Updated: 2023/03/18 03:47:38 by itan             ###   ########.fr       */
+/*   Updated: 2023/03/19 00:54:22 by itan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-long int	get_time_diff(struct timeval start, struct timeval end)
+int	get_time_diff(struct timeval start, struct timeval end)
 {
-	return ((end.tv_sec - start.tv_sec) * 1000 + (end.tv_usec - start.tv_usec)
-		/ 1000);
+	return (int)((end.tv_sec - start.tv_sec) * 1000) + (int)((end.tv_usec
+			- start.tv_usec) / 1000);
 }
